backend/requirements.txt
django == 2.2.1
djangorestframework == 3.10.3
django-rest-knox == 4.1.0
psycopg2 == 2.8.6